+++ src/i18n/en.json
@@ -539,6 +539,7 @@
       "otherGES": "Other gases (in kgCO₂e)",
       "part": "Part",
       "post": "Post",
+      "posts": "Posts",
       "missingPost": "Please select a post",
       "subPost": "Sub Post",
       "partType": "Part type",
